feat(comments): add mention and link formatting in text editors
Add automatic formatting for user mentions (starting with @) and markdown
links in comment fields. Mentions are highlighted in the link color, while
markdown link text is underlined. Raw URLs are also highlighted.

The new `format_comment_on_change()` function uses regex patterns to identify
and format these elements in real-time as users type. This function is
connected to the `contentsChanged` signal in both `AYCommentField` and
`AYTextEditor` components.

Also fix a minor issue in `AYTextBox._build_lower_bar()` where the layout
was incorrectly initialized with `self` as the first argument.

Signed-off-by: philippe-ynput <philippe@ynput.io>